Allow multiple randomly-selected songs for an event
Multiple song names can be provided for each event by separating the
song names with a comma ',' in the properties file. When a song for the
event should load, it will randomly load one of the songs but it will
not play the same song twice in a row. NOTE: This commit introduces a
new song name limitation that song names should not contain the exact
names of any other song names as this will cause conflicts.